CXP-1113: Move conection commads
